# openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Project","description":"プロジェクト管理API"},{"name":"Todo","description":"TODO管理API"}],"paths":{"/api/todos/{id}":{"get":{"tags":["Todo"],"summary":"TODO取得","description":"指定IDのTODOを取得する","operationId":"getById","parameters":[{"name":"id","in":"path","description":"TODO ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"TODOが存在しない","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TodoResponse"}}}},"200":{"description":"取得成功","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TodoResponse"}}}}}},"put":{"tags":["Todo"],"summary":"TODO更新","description":"指定IDのTODOを更新する","operationId":"update","parameters":[{"name":"id","in":"path","description":"TODO ID","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoRequest"}}},"required":true},"responses":{"404":{"description":"TODOが存在しない","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TodoResponse"}}}},"400":{"description":"バリデーションエラー","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TodoResponse"}}}},"200":{"description":"更新成功","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TodoResponse"}}}}}},"delete":{"tags":["Todo"],"summary":"TODO削除","description":"指定IDのTODOを削除する","operationId":"delete","parameters":[{"name":"id","in":"path","description":"TODO ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"削除成功"},"404":{"description":"TODOが存在しない"}}}},"/api/todos":{"get":{"tags":["Todo"],"summary":"TODO一覧取得","description":"登録されている全TODOを取得する","operationId":"getAll","responses":{"200":{"description":"取得成功","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TodoResponse"}}}}}}},"post":{"tags":["Todo"],"summary":"TODO作成","description":"新しいTODOを作成する","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoRequest"}}},"required":true},"responses":{"201":{"description":"作成成功","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TodoResponse"}}}},"400":{"description":"バリデーションエラー","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TodoResponse"}}}}}}},"/api/projects":{"get":{"tags":["Project"],"summary":"プロジェクト一覧取得","operationId":"getAll_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProjectResponse"}}}}}}},"post":{"tags":["Project"],"summary":"プロジェクト作成","operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProjectResponse"}}}}}}},"/api/projects/{id}/todos":{"post":{"tags":["Project"],"summary":"TODO追加","description":"プロジェクトにTODOを追加する","operationId":"addTodo","parameters":[{"name":"id","in":"path","description":"プロジェクトID","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TodoResponse"}}}}}}},"/api/todos/{id}/toggle":{"patch":{"tags":["Todo"],"summary":"完了状態トグル","description":"指定IDのTODOの完了/未完了を切り替える","operationId":"toggleComplete","parameters":[{"name":"id","in":"path","description":"TODO ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"TODOが存在しない","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TodoResponse"}}}},"200":{"description":"更新成功","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TodoResponse"}}}}}}},"/api/projects/{id}/todos/{todoId}/toggle":{"patch":{"tags":["Project"],"summary":"TODO完了トグル","description":"TODOの完了/未完了を切り替える","operationId":"toggleTodo","parameters":[{"name":"id","in":"path","description":"プロジェクトID","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"todoId","in":"path","description":"TODO ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TodoResponse"}}}}}}},"/api/projects/{id}":{"get":{"tags":["Project"],"summary":"プロジェクト取得","operationId":"getById_1","parameters":[{"name":"id","in":"path","description":"プロジェクトID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProjectResponse"}}}}}},"delete":{"tags":["Project"],"summary":"プロジェクト削除","operationId":"delete_1","parameters":[{"name":"id","in":"path","description":"プロジェクトID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/projects/{id}/todos/{todoId}":{"delete":{"tags":["Project"],"summary":"TODO削除","description":"プロジェクトからTODOを削除する","operationId":"removeTodo","parameters":[{"name":"id","in":"path","description":"プロジェクトID","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"todoId","in":"path","description":"TODO ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"TodoRequest":{"required":["title"],"type":"object","properties":{"title":{"maxLength":200,"minLength":0,"type":"string","description":"タイトル","example":"買い物をする"},"description":{"maxLength":1000,"minLength":0,"type":"string","description":"説明","example":"牛乳・卵・パンを買う"},"completed":{"type":"boolean","description":"完了フラグ","example":false}},"description":"TODO作成・更新リクエスト"},"TodoResponse":{"type":"object","properties":{"id":{"type":"integer","description":"ID","format":"int64","example":1},"title":{"type":"string","description":"タイトル","example":"買い物をする"},"description":{"type":"string","description":"説明","example":"牛乳・卵・パンを買う"},"completed":{"type":"boolean","description":"完了フラグ","example":false},"createdAt":{"type":"string","description":"作成日時","format":"date-time"},"updatedAt":{"type":"string","description":"更新日時","format":"date-time"}},"description":"TODOレスポンス"},"ProjectRequest":{"required":["name"],"type":"object","properties":{"name":{"maxLength":100,"minLength":0,"type":"string","description":"プロジェクト名","example":"個人タスク"}},"description":"プロジェクト作成・更新リクエスト"},"ProjectResponse":{"type":"object","properties":{"id":{"type":"integer","description":"ID","format":"int64","example":1},"name":{"type":"string","description":"プロジェクト名","example":"個人タスク"},"todos":{"type":"array","description":"TODO一覧","items":{"$ref":"#/components/schemas/TodoResponse"}}},"description":"プロジェクトレスポンス"}}}}